requirements.txt
pyyaml
apscheduler
pyperclip
requests
pytypedstream
regex
pydantic